refactor engine.py, split into smaller files
